feat: reorganize benchmarks
