+++ crates/algorithms/benches/shortest_paths/dijkstra.rs
@@ -1,12 +1,10 @@
-mod common;
-
 use criterion::{criterion_group, criterion_main, BatchSize, BenchmarkId, Criterion};
-use petgraph_algorithms::shortest_paths::bellman_ford;
+use petgraph_algorithms::shortest_paths::dijkstra;
 
 use crate::common::{nodes, Profile};
 
 fn sparse(criterion: &mut Criterion) {
-    let mut group = criterion.benchmark_group("bellman_ford/sparse");
+    let mut group = criterion.benchmark_group("dijkstra/sparse");
 
     for nodes in nodes(None) {
         group.bench_with_input(
@@ -15,13 +13,13 @@ fn sparse(criterion: &mut Criterion) {
             |bench, &nodes| {
                 bench.iter_batched(
                     || {
-                        let graph = Profile::Sparse.newman_watts_strogatz::<(), f32>(nodes);
+                        let graph = Profile::Sparse.newman_watts_strogatz::<(), u8>(nodes);
                         let nodes = graph.node_indices().collect::<Vec<_>>();
 
                         (graph, nodes)
                     },
                     |(graph, nodes)| {
-                        let _scores = bellman_ford(&graph, nodes[0]);
+                        let _scores = dijkstra(&graph, nodes[0], None, |e| *e.weight());
                     },
                     BatchSize::SmallInput,
                 );
@@ -31,22 +29,22 @@ fn sparse(criterion: &mut Criterion) {
 }
 
 fn dense(criterion: &mut Criterion) {
-    let mut group = criterion.benchmark_group("bellman_ford/dense");
+    let mut group = criterion.benchmark_group("dijkstra/dense");
 
-    for nodes in nodes(Some(128)) {
+    for nodes in nodes(Some(1024)) {
         group.bench_with_input(
             BenchmarkId::from_parameter(*nodes),
             nodes,
             |bench, &nodes| {
                 bench.iter_batched(
                     || {
-                        let graph = Profile::Dense.newman_watts_strogatz::<(), f32>(nodes);
+                        let graph = Profile::Dense.newman_watts_strogatz::<(), u8>(nodes);
                         let nodes = graph.node_indices().collect::<Vec<_>>();
 
                         (graph, nodes)
                     },
                     |(graph, nodes)| {
-                        let _scores = bellman_ford(&graph, nodes[0]);
+                        let _scores = dijkstra(&graph, nodes[0], None, |e| *e.weight());
                     },
                     BatchSize::SmallInput,
                 );
@@ -56,4 +54,3 @@ fn dense(criterion: &mut Criterion) {
 }
 
 criterion_group!(benches, sparse, dense);
-criterion_main!(benches);
