+++ crates/algorithms/src/tree/mod.rs
@@ -1,3 +1,5 @@
-mod min_spanning_tree;
+mod minimum_spanning_tree;
 
-pub use min_spanning_tree::min_spanning_tree;
+#[deprecated(note = "Use `minimum_spanning_tree` instead", since = "0.7.0")]
+pub use minimum_spanning_tree::minimum_spanning_tree as min_spanning_tree;
+pub use minimum_spanning_tree::minimum_spanning_tree;
